requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==3.2
django-allauth==0.50.0
django-braces==1.15.0
django-dotenv==1.4.2
django-widget-tweaks==1.4.12
idna==3.3
mysqlclient==2.1.0
oauthlib==3.2.0
Pillow==9.1.0
pkg_resources==0.0.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.9
